requirements.txt
aiofiles==24.1.0
aiohappyeyeballs==2.6.1
aiohttp==3.11.18
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.9.0
appnope==0.1.4
asttokens==3.0.0
attrs==25.3.0
backoff==2.2.1
beautifulsoup4==4.13.4
certifi==2024.2.2
cffi==1.17.1
chardet==5.2.0
charset-normalizer==3.3.2
click==8.1.8
coloredlogs==15.0.1
comm==0.2.2
contourpy==1.3.2
cryptography==44.0.2
customtkinter==5.2.2
cycler==0.12.1
darkdetect==0.8.0
dataclasses-json==0.6.7
debugpy==1.8.11
decorator==5.1.1
diffusers==0.27.2
distro==1.9.0
docx2txt==0.9
emoji==2.14.1
eval_type_backport==0.2.2
executing==2.1.0
faiss-cpu==1.10.0
fastapi==0.115.12
feedparser==6.0.10
filelock==3.13.1
filetype==1.2.0
flatbuffers==25.2.10
fonttools==4.57.0
frozenlist==1.6.0
fsspec==2024.3.1
groq==0.23.0
h11==0.14.0
html5lib==1.1
httpcore==1.0.8
httpx==0.28.1
httpx-sse==0.4.0
huggingface-hub==0.30.2
humanfriendly==10.0
idna==3.7
importlib_metadata==7.1.0
ipykernel==6.29.5
ipython==8.31.0
jedi==0.19.2
Jinja2==3.1.3
joblib==1.4.2
jsonpatch==1.33
jsonpointer==3.0.0
jupyter_client==8.6.3
jupyter_core==5.7.2
kiwisolver==1.4.8
langchain==0.3.24
langchain-community==0.3.22
langchain-core==0.3.55
langchain-groq==0.3.2
langchain-text-splitters==0.3.8
langdetect==1.0.9
langsmith==0.3.33
lxml==5.4.0
MarkupSafe==2.1.5
marshmallow==3.26.1
matplotlib==3.10.1
matplotlib-inline==0.1.7
mpmath==1.3.0
multidict==6.4.3
mypy_extensions==1.1.0
nest-asyncio==1.6.0
networkx==3.2.1
nltk==3.9.1
numpy==1.26.4
olefile==0.47
onnx==1.17.0
onnxruntime==1.21.1
opencv-python==4.11.0.86
orjson==3.10.16
packaging==24.0
pandas==2.2.3
parso==0.8.4
pdfminer.six==20250416
pexpect==4.9.0
pillow==10.2.0
platformdirs==4.3.6
prompt_toolkit==3.0.48
propcache==0.3.1
protobuf==6.30.2
psutil==6.1.1
ptyprocess==0.7.0
pure_eval==0.2.3
pycparser==2.22
pydantic==2.11.3
pydantic-settings==2.9.1
pydantic_core==2.33.1
Pygments==2.18.0
PyMuPDF==1.25.5
pyparsing==3.2.3
pypdf==5.4.0
PyPDF2==3.0.1
pypdfium2==4.30.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
python-iso639==2025.2.18
python-magic==0.4.27
python-multipart==0.0.20
python-oxmsg==0.0.2
pytz==2025.2
PyYAML==6.0.1
pyzmq==26.2.0
RapidFuzz==3.13.0
regex==2023.12.25
requests==2.31.0
requests-toolbelt==1.0.0
safetensors==0.5.3
scikit-learn==1.6.1
scipy==1.15.2
sentence-transformers==4.1.0
six==1.17.0
sniffio==1.3.1
soupsieve==2.7
SQLAlchemy==2.0.40
stack-data==0.6.3
starlette==0.46.2
sympy==1.12
tenacity==9.1.2
threadpoolctl==3.6.0
timm==1.0.15
tokenizers==0.21.1
torch==2.2.1
torchvision==0.17.1
tornado==6.4.2
tqdm==4.66.2
traitlets==5.14.3
transformers==4.51.3
typing-inspect==0.9.0
typing-inspection==0.4.0
typing_extensions==4.13.2
tzdata==2025.2
unstructured==0.17.2
unstructured-client==0.34.0
unstructured-inference==0.8.10
urllib3==2.2.1
uvicorn==0.34.2
wcwidth==0.2.13
webencodings==0.5.1
wrapt==1.17.2
yarl==1.20.0
zipp==3.18.1
zstandard==0.23.0
newsapi-python==0.2.7
faiss-cpu
duckduckgo-search
apscheduler
schedule
bs4
# JWT and Authentication dependencies
python-jose==3.3.0
passlib==1.7.4
bcrypt==4.1.2
email-validator==2.1.1
# PostgreSQL database adapter
psycopg2-binary==2.9.9
# Database migrations tool
alembic==1.13.1